+++ src/components/Layout/Layout.js
@@ -6,6 +6,15 @@ import {
   withRouter,
 } from "react-router-dom";
 import classnames from "classnames";
+import {Box, IconButton, Link} from '@material-ui/core'
+import Icon from '@mdi/react'
+
+//icons
+import {
+  mdiFacebook as FacebookIcon,
+  mdiTwitter as TwitterIcon,
+  mdiGithub as GithubIcon,
+} from '@mdi/js'
 
 // styles
 import useStyles from "./styles";
@@ -57,6 +66,81 @@ function Layout(props) {
               <Route path="/app/ui/icons" component={Icons} />
               <Route path="/app/ui/charts" component={Charts} />
             </Switch>
+            <Box
+              mt={5}
+              width={"100%"}
+              display={"flex"}
+              alignItems={"center"}
+              justifyContent="space-between"
+            >
+              <div>
+                <Link
+                  color={'primary'}
+                  href={'https://flatlogic.com/'}
+                  target={'_blank'}
+                  className={classes.link}
+                >
+                  Flatlogic
+                </Link>
+                <Link
+                  color={'primary'}
+                  href={'https://flatlogic.com/about'}
+                  target={'_blank'}
+                  className={classes.link}
+                >
+                  About Us
+                </Link>
+                <Link
+                  color={'primary'}
+                  href={'https://flatlogic.com/blog'}
+                  target={'_blank'}
+                  className={classes.link}
+                >
+                  Blog
+                </Link>
+              </div>
+              <div>
+                <Link
+                  href={'https://www.facebook.com/flatlogic'}
+                  target={'_blank'}
+                >
+                  <IconButton aria-label="facebook">
+                    <Icon
+                      path={FacebookIcon}
+                      size={1}
+                      color="#6E6E6E99"
+                    />
+                  </IconButton>
+                </Link>
+                <Link
+                  href={'https://twitter.com/flatlogic'}
+                  target={'_blank'}
+                >
+                  <IconButton aria-label="twitter">
+                    <Icon
+                      path={TwitterIcon}
+                      size={1}
+                      color="#6E6E6E99"
+                    />
+                  </IconButton>
+                </Link>
+                <Link
+                  href={'https://github.com/flatlogic'}
+                  target={'_blank'}
+                >
+                  <IconButton
+                    aria-label="github"
+                    style={{marginRight: -12}}
+                  >
+                    <Icon
+                      path={GithubIcon}
+                      size={1}
+                      color="#6E6E6E99"
+                    />
+                  </IconButton>
+                </Link>
+              </div>
+            </Box>
           </div>
         </>
     </div>