+++ src/components/PageTitle/PageTitle.js
@@ -1,5 +1,4 @@
 import React from "react";
-import { Button } from "@material-ui/core";
 
 // styles
 import useStyles from "./styles";
@@ -15,16 +14,7 @@ export default function PageTitle(props) {
       <Typography className={classes.typo} variant="h1" size="sm">
         {props.title}
       </Typography>
-      {props.button && (
-        <Button
-          classes={{ root: classes.button }}
-          variant="contained"
-          size="large"
-          color="secondary"
-        >
-          {props.button}
-        </Button>
-      )}
+      {props.button && props.button}
     </div>
   );
 }