+++ src/themes/default.js
@@ -120,6 +120,7 @@ export default {
     MuiTableCell: {
       root: {
         borderBottom: "1px solid rgba(224, 224, 224, .5)",
+        paddingLeft: 24
       },
       head: {
         fontSize: "0.95rem",
@@ -128,5 +129,10 @@ export default {
         fontSize: "0.95rem",
       },
     },
+    PrivateSwitchBase: {
+      root: {
+        marginLeft: 10
+      }
+    }
   },
 };